Safari-609.1.11.1/LayoutTests/fast/viewport/viewport-134-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "device-width;" for key "width" is invalid, and has been ignored. Note that ';' is not a separator in viewport values. The list should be comma-separated.
ALERT: viewport size 160x176 scale 2 with limits [2, 5] and userScalable true
